\name{poisson-package}
\alias{poisson-package}
\alias{poisson}
\docType{package}
\title{
Simulating Homogenous & Non-Homogenous Poisson Processes
}
\description{
A collection of functions for simulating homogenous and non-homogenous poisson processes
}
\details{
\tabular{ll}{
Package: \tab poisson\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-10-01\cr
License: \tab GPL-2\cr
}
% An overview of how to use the package, including the most important functions ~~
}
\author{
Kristian Brock [aut], Daniel Slade [ctb]

Maintainer: Daniel Slade <sladeD@bham.ac.uk>
}

